build: reorganize osal and observers include path.
