+++ include/espFoC/esp_foc.h
@@ -35,8 +35,8 @@
 #include "espFoC/drivers/inverter_interface.h"
 #include "espFoC/drivers/current_sensor_interface.h"
 #include "espFoC/drivers/rotor_sensor_interface.h"
-#include "espFoC/drivers/os_interface.h"
-#include "espFoC/esp_foc_observer_interface.h"
+#include "espFoC/osal/os_interface.h"
+#include "espFoC/observer/esp_foc_observer_interface.h"
 #include "espFoC/esp_foc_units.h"
 
 typedef enum {
